work on marching cubes

+++ 03_marching_cubes/simple_mesh.h
@@ -8,7 +8,7 @@
 
 using namespace cv;
 
-typedef Vec3f Vertex;
+typedef Vec3d Vertex;
 
 struct MeshTriangle
 {
